+++ R/paRaoP.R
@@ -17,26 +17,30 @@
 #' @param isfloat Logical; indicates whether the input data values are floating-point numbers.
 #' @param mfactor Integer; multiplication factor in case of input data as float numbers.
 #' @param np Number of processes for parallel computation.
+#' @param progBar logical. If TRUE a progress bar is shown.
 #' @return A list of matrices corresponding to the computed Rao's index values. Each matrix in the list 
 #' represents the calculations performed over the moving window, with dimensions equal to \code{dim(x)}.
 #' @author Duccio Rocchini \email{duccio.rocchini@@unibo.it},
 #' Matteo Marcantonio \email{marcantoniomatteo@@gmail.com}
 #' @seealso \code{\link{paRao}} for the related non-parallelized function.
 
-paRaoP <- function(x,alpha,window,dist_m,na.tolerance,diag,debugging,isfloat,mfactor,np) {
+paRaoP <- function(x,alpha,window,dist_m,na.tolerance,diag,debugging,isfloat,mfactor,np,progBar) {
 # `win` is the operative moving window
 win = window 
 NAwin <- 2*window+1
 
 message("\n\nProcessing alpha: ",alpha, " Moving Window: ", NAwin)
   # Set a progress bar
-  pb <- progress::progress_bar$new(
-    format = "[:bar] :percent in :elapsed\n",
-    # Total number of ticks is the number of column +NA columns divided the number of processor.
-    total = (dim(x)[2]/np)+(dim(x)[2]*0.05), 
-    clear = FALSE, 
-    width = 60, 
-    force = FALSE)
+    if(progBar) {
+        pb <- progress::progress_bar$new(
+            format = "[:bar] :percent in :elapsed\n",
+            # Total number of ticks is the number of column +NA columns divided the number of processor.
+            total = (dim(x)[2]/np)+(dim(x)[2]*0.05), 
+            clear = FALSE, 
+            width = 60, 
+            force = FALSE)
+    }
+    
     # Some initial housekeeping
     mfactor <- ifelse(isfloat,mfactor,1) 
     diagonal <- ifelse(diag==TRUE,0,NA)
@@ -67,7 +71,7 @@ message("\n\nProcessing alpha: ",alpha, " Moving Window: ", NAwin)
      out <- foreach::foreach(cl=(1+win):(dim(x)[2]+win),.verbose = F) %dopar% {
         if(debugging) {cat(paste(cl))}
     # Update progress bar
-    pb$tick()
+    if(progBar) pb$tick()
     # Row loop
     paRaoOP <- sapply((1+win):(dim(x)[1]+win), function(rw) {
         if( length(!which(!tx[c(rw-win):c(rw+win),c(cl-win):c(cl+win)]%in%NA)) < floor(NAwin^2-((NAwin^2)*na.tolerance) ) ) {
@@ -125,10 +129,10 @@ return(do.call(cbind,out))
             cat(paste(cl))
         }
     # Update progress bar
-    pb$tick()
+    if(progBar) pb$tick()
     # Row loop
     paRaoOP <- sapply((1+win):(dim(x)[1]+win), function(rw) {
-        if( length(!which(!tx[c(rw-win):c(rw+win),c(cl-win):c(cl+win)]%in%NA)) < floor(NAwim^2-((NAwim^2)*na.tolerance)) ) {
+        if( length(!which(!tx[c(rw-win):c(rw+win),c(cl-win):c(cl+win)]%in%NA)) < floor(NAwin^2-((NAwin^2)*na.tolerance)) ) {
             vv <- NA
             return(vv)
             }else{
@@ -185,7 +189,7 @@ return(do.call(cbind,out))
                 cat(paste(cl))
             }
         # Update progress bar
-        pb$tick()
+        if(progBar) pb$tick()
         # Row loop
         paRaoOP <- sapply((1+win):(dim(x)[1]+win), function(rw) {
             if( length(!which(!tx[c(rw-win):c(rw+win),c(cl-win):c(cl+win)]%in%NA)) <= (window^2-((window^2)*na.tolerance)) ) {