+++ R/load_copNDVI.R
@@ -8,9 +8,7 @@ load_copNDVI <- function() {
   # Set the path to the RDS file
   rds_file <- system.file("extdata", "copNDVI.rds", package = "rasterdiv")
   
-  # Check if the file exists
   if (file.exists(rds_file)) {
-    # Load the SpatRaster object from the RDS file
     copNDVI <- terra::rast(readRDS(rds_file))
   } else {
     stop("Data file 'copNDVI.rds' not found in the package.")