+++ NEWS.md
@@ -1,6 +1,7 @@
 # rasterdiv 0.3.5
 ## Major changes
 * Adds `twdtw` as a distance methods for `paRao()`
+* Adds back `Rao()` as an alias for `paRao(alpha=2)`
 
 ## Minor changes
 * Changes how `N` is intended in `mpaRaoAreaS`